% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yandex.R
\name{get_translation_direction}
\alias{get_translation_direction}
\title{Gets a list of translation directions supported by the service}
\usage{
get_translation_direction(api_key, lang = "")
}
\arguments{
\item{api_key}{yandex API key}

\item{lang}{If set, the response contains explanations of language codes. Language names are output in the language corresponding to the code in this parameter.}
}
\value{
data frame giving supported translation direction
}
\description{
Gets a list of translation directions supported by the service
}
\examples{
data=get_translation_direction(api_key)
}

